implemented radial damage for rockets, stuns changed enemy movement to use douvbles

+++ src/gameObjects/Enemy.java
@@ -7,8 +7,8 @@ import java.awt.*;
 
 public class Enemy extends GameObject{
 //	private Coordinate pos;
-	private int pathIndex;
-	private int speed;
+	private double pathIndex;
+	private double speed;
 	private double health;
 	private int type;
 	private int width, height;
@@ -52,13 +52,13 @@ public class Enemy extends GameObject{
 		health -= i;
 	}
 
-	public int getPathIndex() {
+	public double getPathIndex() {
 		return pathIndex;
 	}
-	public int getSpeed() {
+	public double getSpeed() {
 		return speed;
 	}
-	public void setPathIndex(int i) {
+	public void setPathIndex(double i) {
 		pathIndex =i;
 	}
 //	public void setPos(Coordinate pos) {
@@ -107,6 +107,8 @@ public class Enemy extends GameObject{
 	public int getHeight() {
 		return height;
 	}
+
+
 //	public Hitbox getHitBox() {
 //		return new Hitbox(pos.getX()-width/2,pos.getY()-height/2,width,height);
 //		return new Rectangle(pos.getX()-width/2,pos.getY()-height/2,width,height);
